Add files via upload
You must be tired using the Function Generator and the CRO during the EEDC Lab. But the Midsem Lab Test is yet to come. You have to prepare for it and perform all the experiments in order to know the working of those machines.

You decide to go to the lab and try those experiments but the lab is locked. The key is with the CC3 guard and he will give you the key only if you answer his queries.

The guard loves the numbers divisible by 2,3 and 5.

Lucky indices are those index for which the sum of numbers formed using the array elements on its left and array elements on its right is divisible by 2,3 and 5.
For example:- Consider the array  and index 3 (1 - based). The number formed using the array elements on the left is  and the number formed using the array elements on the right is .

So he gives you an array of integers and asks you Q queries. In each query he provides you two integers L and R. You have to answer how many lucky indices are there between L to R (both inclusive).

See sample for more clarity.

Input Format:

First line contains an integer N, which denotes the length of the array A.
Second line contains N space-separated integers.
Next line contains number of queries Q.
Next Q lines contain two integers L and R.

Output Format

For each query, print the number of lucky indices present in that range.

Constraints

1 ≤  ≤ 6


Problem Setter - Harsh Jain

SAMPLE INPUT 
7
1 5 7 5 6 7 5
5
1 2
2 3
4 6
1 7
3 6
SAMPLE OUTPUT 
0
0
1
1
1
Explanation
For
Index 1: Left Number: 0 and Right Number: 
Index 2: Left Number: 1 and Right Number: 
Index 3: Left Number:  and Right Number: 
Index 4: Left Number:  and Right Number: 
Index 5: Left Number:  and Right Number: 
Index 6: Left Number:  and Right Number: 5
Index 7: Left Number:  and Right Number: 0
Only sum of numbers at index 5 is divisible by 2,3 and 5.