Add files via upload
Monk and circular distance
Its time for yet another challenge, and this time it has been prepared by none other than Monk himself for Super-Hardworking Programmers like you. So, this is how it goes:

Given N points located on the co-ordinate plane, where the  point is located at co-ordinate , , you need to answer q queries.

In the  query, you shall be given an integer , and considering you draw a circle centered at the origin  with radius , you need to report the number of points lying inside or on the circumference of this circle.

For each query, you need to print the answer on a new line.

Input Format :

The first line contains a single integer N denoting the number of points lying on the co-ordinate plane. Each of the next N lines contains 2 space separated integers  and , denoting the x and y co-ordintaes of the  point.

The next line contains a single integer q, denoting the number of queries. Each of the next q lines contains a single integer, where the integer on the  line denotes the parameters of the  query .

Output Format :

For each query, print the answer on a new line.

Constraints :





Sample Input
5
1 1
2 2
3 3
-1 -1
4 4
2
3
32
Sample Output
3
5
Explanation
For the  query in the sample, the circle with radius equal to 3 looks like this on the co-ordinate plane

enter image description here

The points : ,  and  lie inside or on the circumference of the circle.